+++ test/test_formatters.py
@@ -37,6 +37,7 @@ TABULAR_RESPONSES = [
     [j_taxon_1, j_taxon_2_partial],
     [j_user_1, j_user_2_partial],
     [j_quality_metric_1],
+    [j_vote_1],
 ]
 
 
@@ -121,6 +122,7 @@ Observation(
         Comment(id=2071896, username='borisb', created_at='Sep 05, 2018', truncated_body='I now see: Bonus species on observation! You ma...'),
         Comment(id=2071611, username='borisb', created_at='Sep 05, 2018', truncated_body='suspect L. bardanae - but sits on Solanum (non-...')
     ],
+    faves=[],
     identifications=[
         Identification(id=34896306, username='niconoe', taxon_name='Genus Lixus', created_at='Sep 05, 2018', truncated_body=''),
         Identification(id=34926789, username='borisb', taxon_name='Lixus bardanae', created_at='Sep 05, 2018', truncated_body=''),
@@ -135,7 +137,8 @@ Observation(
     quality_metrics=[],
     sounds=[],
     taxon=Taxon(id=493595, full_name='Lixus bardanae'),
-    user=User(id=886482, login='niconoe', name='Nicolas Noé')
+    user=User(id=886482, login='niconoe', name='Nicolas Noé'),
+    votes=[]
 )
 """